moved stores to stores folder
