+++ src/stores/settings.ts
